Wait for approval before continuing donation
Before, the donation flow did not wait for the approval transaction
to be mined before restarting the donation. This created a race
condition where it would continually prompt the user to approve
until the first approval transaction finally got mined. This commit
fixes this by waiting until the approval tx is mined before checking
the allowance again and continuing the rest of the donation flow.