Extract donation flow into a wrapper component
The `withDonationFlow` component wraps the form and modal and handles
the state of the actual donation, passing information about its current
step to the wrapped component as props. If something goes wrong, it
throws an error and lets the wrapped component handle it.

Wrap the `Donation` and `Sponsorship` components and leave just the
data. Convert them to simple function components.